Reformat command registry, move ArgumentTypeWrapper for clarity
